mDPO/requirements.txt
# conda create --name mdpo python=3.10
# conda activate mdpo
torch==2.0.1
torchvision==0.15.2
numpy==1.26.2
accelerate==0.27.2
datasets==2.14.6
transformers==4.38.0
trl==0.7.2
peft==0.5.0
deepspeed==0.11.0
PyYAML==6.0.1
submitit==1.5.1
spacy==3.6.0
nltk==3.8.1
# python -m spacy download en_core_web_sm
# python -m spacy download en_core_web_lg
# python -c "import nltk; nltk.download('punkt')"
# python -c "import nltk; nltk.download('averaged_perceptron_tagger')"
# python -c "import nltk; nltk.download('wordnet')"
einops==0.8.1
tiktoken==0.9.0
matplotlib==3.10.1
pillow==11.2.1
transformers_stream_generator
wandb==0.19.9
